% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioturbateTimeseries.R
\name{BioturbateTimeseries}
\alias{BioturbateTimeseries}
\title{BioturbateTimeseries}
\usage{
BioturbateTimeseries(tss, timepoints, bio_depth = 0.1, acc_rate = 0.01,
  biowidth_timesteps, n_samples = Inf, return_type = c("list", "data.frame",
  "ts"))
}
\arguments{
\item{tss}{a ts object containing an equidistant timeseries, (e.g. model
result or interpolated time series)}

\item{timepoints}{the timepoints at which to sample the time series}

\item{bio_depth}{depth of the bioturbated layer in metres, defaults to 0.1 m}

\item{acc_rate}{rate of sediment accumulation (sedimentation rate) in metres
per year. Defaults to 0.01 m (1 cm per year).}

\item{n_samples}{Number of e.g. foraminifera sampled per timepoint}

\item{return_type}{Should the function return a data.frame, list or time series. Defaults to list.}
}
\value{
a list containing the resampled data values and times
}
\description{
Resamples an equidistant sampled timeseries (e.g. model result)
  at the specified "timepoints" by simulating the foraminifera sampling
}
\examples{
tss <- ts(c(rep(0, 100), 1, rep(0, 99)), 0:199))
timepoints <- c(time(tss))
par(mfcol = c(1, 2))
plot(tss)
biowidth_timesteps = 10
result <- BioturbateTimeseries(tss, timepoints, biowidth_timesteps, n_samples = 100)
plot(result$time, result$data, type = "l")
par(mfcol = c(1, 1))

tss<-ts(sin(seq(-pi, 9*pi, length.out = 1000)), start = 0, frequency = 1/100)
timepoints<-c(time(tss)[1:1000])
biowidth_timesteps=100
result<-BioturbateTimeseries(tss,timepoints,biowidth_timesteps, n_samples = 30)

plot(tss)
lines(data~time, data = result, col = "Red")
}

