% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioturbationWeights.R
\name{BioturbationWeights}
\alias{BioturbationWeights}
\title{Title}
\usage{
BioturbationWeights(z = NULL, focal.depth, layer.width, sed.acc.rate,
  mix.depth)
}
\arguments{
\item{mix.depth}{}
}
\description{
Title
}
\examples{
z <- 0:10000
w <- BioturbationWeights(z, focal.depth = 2000, layer.width = 1, sed.acc.rate = 5 / 1000, mix.depth = 10)
plot(z, w, "l")
}
