% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-uncertainty.R
\name{CalErr}
\alias{CalErr}
\title{Calibration (regression) error from RSE and x values of fitted regression}
\usage{
CalErr(x_star, rse, x)
}
\arguments{
\item{x_star}{New x values for which error in y is desired}

\item{rse}{Residual Standard Error of fitted regression}

\item{x}{x values from fitted regression}
}
\value{
vector
}
\description{
Calibration (regression) error from RSE and x values of fitted regression
}
\examples{
n <- 20
df <- data.frame(x = runif(n, 50, 100))
df$y <- 1 + 0.5 * df$x + rnorm(n, 0, 10)

plot(df)
lm1 <- lm(y~x, data = df)
abline(lm1)

####

library(dplyr)
library(ggplot2)

n <- 20
df <- data.frame(x = runif(n, 50, 100))
df$y <- 1 + 0.5 * df$x + rnorm(n, 0, 10)

plot(df)
lm1 <- lm(y~x, data = df)
abline(lm1)

CalErr(c(23, 56, 79), summary(lm1)$sigma, df$x)

df2 <- tibble(x = 0:200,
              y = predict(lm1, newdata = data.frame(x=x))) \%>\%
  mutate(y.err = CalErr(x_star = x, rse = summary(lm1)$sigma, x = df$x))

df \%>\%
  ggplot(aes(x , y)) +
  geom_point() +
  geom_smooth(method = "lm") +
  geom_line(data = df2, aes(x, y = y + 2*y.err))+
  geom_line(data = df2, aes(x, y = y - 2*y.err))

}
