% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.uk37.R
\name{calib.uk37}
\alias{calib.uk37}
\title{Convert between UK'37 and Temperature in degrees C}
\usage{
calib.uk37(temperature = NULL, UK37 = NULL, point.or.sample = c("point",
  "sample"), n = 1)
}
\arguments{
\item{n}{}
}
\description{
Convert between UK'37 and Temperature in degrees C
}
