% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.functions.R
\name{CalibMgCa}
\alias{CalibMgCa}
\title{Convert between MgCa and Temperature in degrees C}
\usage{
CalibMgCa(temperature = NULL, proxy.value = NULL,
  point.or.sample = c("point", "sample"), n = 1,
  slp.int.means = NULL, slp.int.vcov = NULL,
  taxon = c("10 Foram Taxa", "G. aequilateralis_350-500",
  "G. aequilateralis_500-1000", "G. conglobatus_350-500",
  "G. hirsuta_350-500", "G. inflata_350-500", "G. ruber pink_250-350",
  "G. ruber pink_350-500", "G. ruber white_250-350",
  "G. ruber white_350-500", "G. sacculifer with sac_350-500",
  "G. sacculifer without sac_350-500", "G. truncatulinoides_350-500",
  "G. truncatulinoides_500-1000", "N. dutertrei_350-500",
  "O. univesa_350-500", "P. obliquiloculata_350-500"))
}
\arguments{
\item{temperature}{Temperature in degrees C}

\item{proxy.value}{Temperature in proxy units}

\item{point.or.sample}{Use the "best estimate" calibration parameters, or
parameters sampled from the fitted calibration model}

\item{n}{the number of replicate conversions to make in the case of sampled
calibration parameters}

\item{slp.int.means}{Optional user supplied vector of values for the slope
and intercept of the calibration function. Overides the defaults.}

\item{slp.int.vcov}{Optional user supplied variance covariance matrix
calibration parameters. Overides the defaults.}

\item{taxon}{The name of a specific taxon for which calibration parameters
are provided by sedproxy. Currently applies only to proxy.calibration.type MgCa.}
}
\value{
a vector of temperatures or proxy values
}
\description{
Convert between MgCa and Temperature in degrees C
}
\examples{
# From temperature to MgCa
## With fixed calibration
CalibMgCa(temperature = c(20, 25), point.or.sample = "point")

## With random calibration, 5 replicates
CalibMgCa(temperature = c(20, 25), n = 5, point.or.sample = "sample")


## Back-transformation with same calibration
CalibMgCa(proxy.value = as.vector(CalibMgCa(temperature = c(21, 22), point.or.sample = "point"))
           , point.or.sample = "point")

## Back-transformation with random calibration
CalibMgCa(proxy.value = as.vector(CalibMgCa(temperature = c(21, 22), point.or.sample = "point"))
           , n = 5, point.or.sample = "sample")

## Incompatible arguments
\dontrun{
CalibMgCa(temperature = 1, proxy.value = 1)
}
}
\seealso{
Other calib: \code{\link{CalibUK37}},
  \code{\link{ProxyConversion}}
}
\concept{calib}
