% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimToProxyClim.dev.R
\name{ClimToProxyClim.dev}
\alias{ClimToProxyClim.dev}
\title{Simulate sediment archived proxy records from an "assumed true" climate signal.}
\usage{
ClimToProxyClim.dev(clim.signal, timepoints,
  proxy.calibration.type = c("identity", "UK37", "MgCa"),
  smoothed.signal.res = 100, habitat.weights = rep(1, 12),
  habitat.wt.args = NULL, bio.depth = 10, sed.acc.rate = 50,
  meas.noise = 0, meas.bias = 0, n.samples = Inf, n.replicates = 1)
}
\arguments{
\item{clim.signal}{The "assumed true" climate signal, e.g. climate model
output or instrumental record. A \code{\link{ts}} object consisting of a
years x 12 (months) x n habitats (e.g. depths) matrix of temperatures. The
time series should be at annual resolution and in reverse, i.e. "most
recent timepoint first" order.}

\item{timepoints}{The timepoints for which the proxy record is to be modelled}

\item{proxy.calibration.type}{Type of proxy, e.g. Uk'37 or MgCa, to which the
clim.signal is converted before the archiving and measurement of the proxy
is simulated. Defaults to "identity" which means no conversion takes place.}

\item{smoothed.signal.res}{The resolution, in years, of the smoothed (block
averaged) version of the input climate signal returned for plotting. This
does not affect what the proxy model uses as input. If set to NA, no
smoothed climate output is generated, this can speed up some simulations.}

\item{habitat.weights}{Either the seasonal pattern of productivity for the organism(s)
recording / producing the proxy as a vector of 12 values, or a function that
produces an index of productivity as a function of temperature.
Defaults to a uniform seasonal distribution.}

\item{habitat.wt.args}{A named list of parameters to be passed to a function named in habitat.weights}

\item{bio.depth}{Depth of the bioturbated layer in cm, defaults to 10 cm.}

\item{sed.acc.rate}{Sediment accumulation rate in cm per 1000 years. Defaults
to 50 cm per ka. Either a single value, or vector of same length as
"timepoints"}

\item{meas.bias}{The amount of bias to add to each simulated proxy
time-series. Each replicate proxy time-series has a constant bias added,
drawn from a normal distribution with mean = 0, sd = meas.bias. Bias
defaults to zero.}

\item{n.samples}{Number of e.g. Foraminifera sampled per timepoint, this can
be either a single number, or a vector of length = timepoints}

\item{n.replicates}{Number of replicate proxy time-series to simulate from
the climate signal}
}
\value{
\code{ClimToProxyClim} returns a list with three elements:
\enumerate{
\item a dataframe \code{simulated.proxy}
\item a dataframe \code{smoothed.signal}
\item a dataframe \code{everything}
}

The dataframe \code{simulated.proxy} contains a single realisation of the
final forward modelled proxy, as well as the intermediate stages and the
original climate signal at the requested timepoints.

The dataframe \code{smoothed.signal} contains a block averaged version the
input climate signal, defaults to 100 year means but this is set by the
parameter smoothed.signal.res. This is useful for plotting against the
resulting simulated proxy.

The dataframe \code{everything} contains all of the above but with multiple
replicates of the pseudo-proxy records if requested. The data are in
"long form", with the column "stage" inidcating the proxy stage or input
climate resolution and column "value" giving the values.

\strong{Named elements of the returned proxy record:}

\describe{
\item{timepoints}{Requested timepoints}
\item{clim.signal.ann}{Input climate signal at requested timepoints at annual resolution}
\item{clim.signal.smoothed}{Input climate signal at regular time intervals and resolution = smoothed.signal.res}
\item{clim.timepoints.ssr}{Input climate signal at requested timepoints, smoothed to resolution = smoothed.signal.res}
\item{proxy.bt}{Climate signal after bioturbation}
\item{proxy.bt.sb}{Climate signal after bioturbation and habitat bias}
\item{proxy.bt.sb.inf.b}{Climate signal after bioturbation, habitat bias, and calibration bias}
\item{proxy.bt.sb.inf.b.n}{Climate signal after bioturbation, habitat bias, and measurement error}
\item{proxy.bt.sb.sampY}{Climate signal after bioturbation, habitat bias, and aliasing of inter-annual variation}
\item{proxy.bt.sb.sampYM}{Climate signal after bioturbation, habitat bias, and aliasing of inter-annual and intra-annual variation such as monthly temperatures or depth habitats}
\item{proxy.bt.sb.sampYM.b}{Climate signal after bioturbation, habitat bias, and aliasing of inter-annual and intra-annual variation such as monthly temperatures or depth habitats, and calibration bias}
\item{proxy.bt.sb.sampYM.b.n}{Climate signal after bioturbation, habitat bias, aliasing, and measurement error}
\item{simulated.proxy}{Final simulated pseudo-proxy, this will be same as proxy.bt.sb.inf.b.n when n.samples = Inf, and proxy.bt.sb.sampYM.b.n when n.samples is finite}
\item{observed.proxy}{True observed proxy (when supplied)}
}
}
\description{
\code{ClimToProxyClim} simulates the creation of a proxy climate
record from a climate signal that is assumed to be true.

The following aspects of proxy creation are currently modelled.
\enumerate{
\item Seasonal bias in the encoding of a proxy due to the interaction between
climate seasonality and any seasonality in the life cycle of the organism
encoding the climate signal (e.g. Foraminifera for Mg/Ca ratios, or
phytoplankton for Alkenone unsaturation indices).
\item Bioturbation of the sediment archived proxy. For each requested
timepoint, the simulated proxy consists of a weighted mean of the climate
signal over a time window that is determined by the sediment accumulation
rate {sed.acc.rate} and the bioturbation depth {bio.depth} which defaults
to 10 cm. The weights are given by the depth solution to an impulse
response function (Berger and Heath, 1968).
\item Aliasing of seasonal and inter-annual climate variation onto to
bioturbated (smoothed) signal. For proxies measured on a small number of
discrete particles both seasonal and inter-annual climate variation is
aliased into the proxy record. For example, Foraminifera have a life-cycle
of approximately 1 month, so they record something like the mean
temperature from a single month. If Mg/Ca is measured on e.g.
\code{n.samples} = 30 individuals, the measured proxy signal is a mean of
30 distinct monthly mean temperatures and will thus be a stochastic sample
of the true mean climate.
\item Measurement noise/error is added as a pure Gaussian white noise process
with mean = 0, standard deviation =
\code{sqrt(sigma.measurement^2 + sigma.individual^2/n.samples)}.
\item Additionally, a random \emph{bias} can be added to each realisation of a
proxy record. Bias is simulated as a Gaussian random variable with mean =
0, standard deviation = \code{meas.bias}. The same randomly generated bias
value is applied to all timepoints in a simulated proxy record, when
multiple replicate proxies are generated ({n.replicates} > 1) each
replicate has a different bias applied.
}

\code{ClimToProxyClim} returns one or more replicates of the final
simulated proxy as well as several intermediate stages (see section
\strong{Value} below).
}
