% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimToProxyClim.R
\name{ClimToProxyClim}
\alias{ClimToProxyClim}
\title{Title}
\usage{
ClimToProxyClim(clim.signal, timepoints, seas.prod, bio.depth = 0.1,
  acc.rate = 0.012, n_samples = Inf)
}
\arguments{
\item{clim.signal}{The "true" climate signal, i.e. model output or instrumental record. A years x 12 (months) matrix of temperatures.}

\item{timepoints}{The timepoints at which the proxy is to be sampled}

\item{seas.prod}{The seasonal pattern of productivity for the organism(s) archived in the proxy. Either a vector of 12 values or a matrix of the same dimensions as clim.signal}

\item{bio.depth}{Depth of the bioturbated layer in metres, defaults to 0.1 m}

\item{acc.rate}{rate of sediment accumulation (sedimentation rate) in metres
per year. Defaults to 0.01 m (1 cm per year).}

\item{n_samples}{Number of e.g. foraminifera sampled per timepoint}
}
\description{
Title
}

