% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimToProxyClim.R
\name{ClimToProxyClim_Rapid}
\alias{ClimToProxyClim_Rapid}
\title{Fast version of ClimToProxyClim.}
\usage{
ClimToProxyClim_Rapid(clim.signal, timepoints, plot.sig.res,
  habitat.weights, bio.depth, sed.acc.rate, layer.width, sigma.meas,
  sigma.ind, meas.bias, n.samples, n.replicates)
}
\description{
Fast version of ClimToProxyClim for simulations with
time-invariant parameters. Used internally by ClimToProxyClim.
}
