% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTaxonTemperature.R
\name{dTaxonTemperature}
\alias{dTaxonTemperature}
\title{Foram production as a function of temperature, using published temperature preference parameters.}
\usage{
dTaxonTemperature(temperature, taxon = taxa.temperature.prefs$ID)
}
\arguments{
\item{temperature}{temperature in degrees celsius}

\item{taxon}{taxon ID in table taxa.temperature.prefs,}
}
\description{
Foram production as a function of temperature, using published temperature preference parameters.
}
