% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\name{example.data}
\alias{example.data}
\title{Data for running the examples in package \code{sedproxy}}
\description{
A set of data objects for running the examples in \code{sedproxy}

\code{\link{N41.proxy}}

\code{\link{N41.proxy.details}}

\code{\link{N41.G.ruber.seasonality}}

\code{\link{N41.t21k.climate}}
}
\details{
Example data
}
