% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPFMs.R
\name{facet_wrap_paginate_auto}
\alias{facet_wrap_paginate_auto}
\title{Plot all ggforce::facet_wrap_paginate pages}
\usage{
facet_wrap_paginate_auto(ggplot.obj, facets, nrow, ncol)
}
\arguments{
\item{ggplot.obj}{A ggplot object}

\item{facets}{Faceting formula, see \code{\link{ggplot2::facet_wrap}}}

\item{nrow}{number of columns per page}

\item{ncol}{number of rows per page}
}
\value{
A list of ggplot objects
}
\description{
A helper function to return all pages from
ggforce::facet_wrap_paginate. It automatically calculates
required number of pages.
}
\examples{
gg <- facet_wrap_paginate_auto(uk37.plots.2,
        facets = formula(~Location + Proxy + ID.no),
        nrow = 4, ncol = 3)
gg
}

