% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimToProxyClim.R
\name{MakePFMDataframe}
\alias{MakePFMDataframe}
\title{Convert "everything" part of output from ClimToProxyClim to dataframe}
\usage{
MakePFMDataframe(PFM)
}
\arguments{
\item{PFM}{output from ClimToProxyClim}
}
\description{
Convert "everything" part of output from ClimToProxyClim to dataframe
}
