% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\docType{data}
\name{N41.G.ruber.seasonality}
\alias{N41.G.ruber.seasonality}
\title{Seasonality of Globigerinoides ruber at core MD97-2141}
\format{
A vector of 12 values
}
\source{
Fraile, I., Schulz, M., Mulitza, S., & Kucera, M. (2008).
Predicting the global distribution of planktonic foraminifera using a
dynamic ecosystem model. Biogeosciences, 5(3), 891–911.
}
\usage{
N41.G.ruber.seasonality
}
\description{
Seasonality of Globigerinoides ruber at core MD97-2141 predicted
by the PLAFOM model (Fraile et al. 2008).
}
\keyword{datasets}
