% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\docType{data}
\name{N41.proxy.details}
\alias{N41.proxy.details}
\title{Metadata for datset \code{N41.proxy}}
\format{
A data frame with 1 rows and 17 variables:
\tabular{ll}{
   \cr \code{Number} \tab Proxy ID number from Shakun et al. (2012)
   \cr \code{ID.no} \tab Proxy ID number from Shakun et al. (2012) with prefix "N"
   \cr \code{Core} \tab ID code of sediment core
   \cr \code{Location} \tab Location of core
   \cr \code{Proxy} \tab Proxy type
   \cr \code{Lat} \tab Latutide of core
   \cr \code{Lon} \tab Longitude of core
   \cr \code{Elevation} \tab Depth of core-top below sea level in metres
   \cr \code{Reference} \tab Original reference for the proxy record
   \cr \code{Resolution} \tab Average time resolution in years
   \cr \code{Calibration.ref} \tab Reference for Mg/Ca calibration
   \cr \code{Calibration} \tab Mg/Ca calibration formula
   \cr \code{Foram.sp} \tab Foram species analyses for proxy
   \cr \code{Ref.14C} \tab Reference for carbon dating
   \cr \code{Notes} \tab
   \cr \code{Geo.cluster} \tab Coarse geographic location
   \cr \code{Archive.type} \tab Type of proxy archive
}
}
\source{
Shakun, J. D., Clark, P. U., He, F., Marcott, S. A.,
Mix, A. C., Liu, Z., … Bard, E. (2012). Global warming preceded by increasing
carbon dioxide concentrations during the last deglaciation. Nature,
484(7392), 49–54. \doi{10.1038/nature10915}
}
\usage{
N41.proxy.details
}
\description{
Metadata for core MD97-2141 taken from Shakun et al. (2012)
}
\keyword{datasets}
