% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\docType{data}
\name{N41.proxy}
\alias{N41.proxy}
\title{Mg/Ca proxy based temperature reconstruction for core MD97-2141}
\format{
A data frame with 216 rows and 4 variables:
\tabular{ll}{
  \cr \code{Published.age} \tab Age in yr BP
  \cr \code{Published.temperature} \tab Reconstructed temperature in degrees C
  \cr \code{Sed.acc.rate.cm.ka} \tab Sediment accumulation rate in cm per kyr
  }
}
\source{
Original reference: Rosenthal, Y., Oppo, D. W., & Linsley, B. K.
(2003). The amplitude and phasing of climate change during the last
deglaciation in the Sulu Sea, western equatorial Pacific. Geophysical
Research Letters, 30(8), 1428. \doi{10.1029/2002GL016612}.
}
\usage{
N41.proxy
}
\description{
Mg/Ca proxy based temperature reconstruction for core MD97-2141, originally published in
Rosenthal et al. (2003), extracted from Shakun et al. (2012).
}
\details{
Published age and published temperature were extracted from
Shakun, J. D., Clark, P. U., He, F., Marcott, S. A.,
Mix, A. C., Liu, Z., … Bard, E. (2012). Global warming preceded by increasing
carbon dioxide concentrations during the last deglaciation. Nature,
484(7392), 49–54. \doi{10.1038/nature10915}.

Sediment accumulation rates were estimated by fitting splines to the published
age and depth estimates.
}
\keyword{datasets}
