% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\docType{data}
\name{N41.t21k.climate}
\alias{N41.t21k.climate}
\title{Climate (surface temperature) at core MD97-2141 from TraCE-21ka}
\format{
A matrix with 22040 rows and 12 columns
}
\source{
Liu, Z., Otto-Bliesner, B. L., He, F., Brady, E. C., Tomas, R.,
Clark, P. U., … Cheng, J. (2009).
Transient Simulation of Last Deglaciation with a New Mechanism for
Bølling-Allerød Warming. Science, 325(5938), 310–314.
https://doi.org/10.1126/science.1171041
}
\usage{
N41.t21k.climate
}
\description{
Modelled surface temperature at core MD97-2141. Model output
from TraCE-21ka simulations.
}
\keyword{datasets}
