% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\docType{data}
\name{param.tab}
\alias{param.tab}
\title{sedproxy parameters}
\format{
A data frame with 12 rows and 4 variables:
\describe{
  \item{\code{Function argument}}{character Argument name in \code{ClimToProxyClim}}
  \item{\code{Description}}{character Description of argument and corresponding variable/parameter}
  \item{\code{Possible sources}}{character Source or possible sources of values for arguments}
  \item{\code{Default}}{character Default values for arguments}
}
}
\usage{
param.tab
}
\description{
Parameters and variables required to generate a pseudo-proxy with \code{ClimToProxyClim}
}
\keyword{datasets}
