% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-uncertainty.R
\name{ScaleError}
\alias{ScaleError}
\title{Title}
\usage{
ScaleError(
  mean.temperature = NULL,
  sd.temperature = NULL,
  mean.proxy.value = NULL,
  sd.proxy.value = NULL,
  calibration.type,
  slp.int.means = NULL,
  slp.int.vcov = NULL,
  calibration = NULL,
  point.or.sample = c("point", "sample")
)
}
\arguments{
\item{point.or.sample}{}
}
\description{
Title
}
