% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-example-data.R
\docType{data}
\name{scussolini.tab1}
\alias{scussolini.tab1}
\title{Scussolini et al. (2013) Table 1}
\format{
A dataframe with 22 rows and 6 columns
}
\source{
Scussolini, P., van Sebille, E., & Durgadoo, J. V. (2013).
Paleo Agulhas rings enter the subtropical gyre during the penultimate
deglaciation. Climate of the Past, 9(6), 2631–2639.
https://doi.org/10.5194/cp-9-2631-2013
}
\usage{
scussolini.tab1
}
\description{
Data from table 1 in Scussolini et al. (2013)
}
\keyword{datasets}
