% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonalityBias.R
\name{SeasonalityBias}
\alias{SeasonalityBias}
\title{Resample a seasonal cycle simulating the seasonal contribution}
\usage{
SeasonalityBias(sst.clim, timepoints, proxy.clim, n_samples = Inf)
}
\arguments{
\item{sst.clim}{ts object containing the monthly SST climatology (12 values)}

\item{timepoints}{the timepoints at which to sample the time series}

\item{proxy.clim}{ts object containing the monthly Foram population index (12 values)}

\item{n_samples}{Number of e.g. foraminifera sampled per timepoint}
}
\value{
A list. result=list(time, data)

data is a single value for each timepoint, indicating the offset in the proxy
temperature record due to the seasonal abundance pattern of the organism,
which preferentially samples those periods of the year when it is most
abundant. There is a stochastic component to this due to measuring a finite
number of individuals. If n_samples is set to \code{Inf}, then these biases will be
constant for a given sst.clim and proxy.clim.
}
\description{
Calculate the (possibly stochastic) bias in the proxy record,
due to seasonality in the abundance of the sampled organism.
}
\examples{

}

