% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-shiny.R
\name{ShinySedproxy}
\alias{ShinySedproxy}
\title{Run ShinySedproxy}
\usage{
ShinySedproxy()
}
\value{
No return value, called for side effects.
}
\description{
Launches the shiny app GUI interface to sedproxy
}
\examples{
\dontrun{ShinySedproxy()}
}
