% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedproxy.R
\docType{data}
\name{stage.labels}
\alias{stage.labels}
\title{Labels for proxy stages}
\format{
A named character vector
}
\usage{
stage.labels
}
\description{
Labels for proxy stages. For plotting.
}
\keyword{datasets}
