% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedproxy.R
\docType{data}
\name{stages.key}
\alias{stages.key}
\title{Description of proxy stages}
\format{
A data frame with 13 rows and 6 variables:
\tabular{ll}{
  \cr  \code{stage}  \tab proxy stage
  \cr  \code{label}  \tab label for proxy stage
  \cr  \code{description}  \tab description of proxy stage
  \cr  \code{plot.order}  \tab default plotting order of stages
  \cr  \code{plotting.colour}  \tab default colour for plotting
  \cr  \code{plotting.alpha}  \tab default alpha level for plotting
}
}
\usage{
stages.key
}
\description{
A description of the proxy stages in the output of \code{ClimToProxyClim}
and default labels, colours and order for plotting
}
\keyword{datasets}
