% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.functions.R
\name{calib.uk37}
\alias{calib.uk37}
\title{Convert between UK'37 and Temperature in degrees C}
\usage{
calib.uk37(temperature = NULL, UK37 = NULL, point.or.sample = c("point",
  "sample"), n = 1)
}
\arguments{
\item{n}{}
}
\description{
Convert between UK'37 and Temperature in degrees C
}
\examples{
# From temperature to UK'37
## With fixed calibration
calib.uk37(temperature = c(1, 2), point.or.sample = "point")

## With random calibration, 5 replicates
calib.uk37(temperature = c(1, 2), n = 5, point.or.sample = "sample")


## Back-transformation with same calibration
calib.uk37(UK37 = as.vector(calib.uk37(temperature = c(21, 22), point.or.sample = "point"))
           , point.or.sample = "point")

## Back-transformation with random calibration
calib.uk37(UK37 = as.vector(calib.uk37(temperature = c(21, 22), point.or.sample = "point"))
           , n = 5, point.or.sample = "sample")

## Incompatible arguments
calib.uk37(temperature = 1, UK37 = 1)
}
