% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.functions.R
\name{CalibrationUncertainty}
\alias{CalibrationUncertainty}
\title{Calibration Uncertainty}
\usage{
CalibrationUncertainty(temperature, means, vcov)
}
\arguments{
\item{temperature}{temperature at which to evaluate proxy uncertainty in
degrees celsius}

\item{means}{intercept and slope of fitted calibration regression}

\item{vcov}{variance-covariance matrix of slope and intercept from fitted
calibration regression}
}
\value{
a tibble (dataframe) with temperature, mean proxy unit and 1 x sigma
  proxy unit
}
\description{
Calculates the calibration uncertainty at a given temperature
  from the variance-covariance matrix of the coefficients of a fitted
  calibration regression equation. Also returns the mean proxy value at the
  given temperature.
}
