% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimToProxyClim.R
\name{ClimToProxyClim}
\alias{ClimToProxyClim}
\title{Simulate sediment archived proxy records from an "assumed true" climate signal.}
\usage{
ClimToProxyClim(clim.signal, timepoints, seas.prod = rep(1, 12),
  bio.depth = 0.1, sed.acc.rate = 5e-04, meas.noise = 0, meas.bias = 0,
  n.samples = Inf, n.replicates = 1)
}
\arguments{
\item{clim.signal}{The "assumed true" climate signal, e.g. climate model output or
instrumental record. A years x 12 (months) matrix of temperatures.}

\item{timepoints}{The timepoints for which the proxy record is to be modelled}

\item{seas.prod}{The seasonal pattern of productivity for the organism(s)
archived in the proxy. Either a vector of 12 values or a matrix of the same
dimensions as clim.signal. Defaults to a uniform seasonal distribution.}

\item{bio.depth}{Depth of the bioturbated layer in metres, defaults to 0.1 m}

\item{sed.acc.rate}{Sediment accumulation rate in metres per year. Defaults
to 5e-04 m per year (0.5 m / kyr). Either a single value, or vector of same
length as "timepoints"}

\item{meas.noise}{The amount of noise to add to each simulated proxy value.
Defined as the standard deviation of a normal distribution with mean = 0}

\item{meas.bias}{The amount of bias to add to each simulated proxy
time-series. Each replicate proxy time-series has a constant bias added,
drawn from a normal distribution with mean = 0, sd = meas.bias. Bias
defaults to zero.}

\item{n.samples}{Number of e.g. Foraminifera sampled per timepoint}

\item{n.replicates}{Number of replicate proxy time-series to simulate from
the climate signal}
}
\value{
\code{ClimToProxyClim} returns a list with three elements:
\enumerate{
\item a dataframe \code{simulated.proxy}
\item a dataframe \code{smoothed.signal}
\item a list \code{everything}
}

The dataframe \code{simulated.proxy} contains a single realisation of the
final forward modelled proxy, as well as the intermediate stages and the
original climate signal at the requested timepoints.

The dataframe \code{smoothed.signal} contains 100 year means of the original climate signal. This
is useful for plotting.

The list \code{everything} contains all of the above, but where a stage contains stochastically
generated noise, rather than a vector, a \code{timepoints} \strong{by} \code{n.replicates} matrix is returned.

\strong{Named elements of the returned proxy record:}

\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
timepoints                 \tab requested timepoints                                                                                                               \cr
clim.timepoints.100 \tab 100 year means of climate signal evaluated at the requested timepoints                                                             \cr
clim.timepoints.50 \tab 50 year means of climate signal evaluated at the requested timepoints                                                             \cr
proxy.bt               \tab proxy after bioturbation                                                                                                           \cr
proxy.bt.sb              \tab proxy after bioturbation + seasonal bias                                                                                           \cr
sed.acc.rate               \tab sediment accumulation rates for each timepoint                                                                                     \cr
smoothing.width            \tab weighted mean time span represented in a sample after bioturbation                                                                 \cr
proxy.bt.sb.sampY             \tab proxy after bioturbation, seasonal bias and finite sampling of years but not seasonality                                                                        \cr
proxy.bt.sb.sampYM             \tab proxy after bioturbation, seasonal bias and finite sampling of years and months                                                                        \cr
proxy.bt.sb.inf.b            \tab proxy after bioturbation, seasonal bias and calibration bias                                                                       \cr
proxy.bt.sb.inf.b.n          \tab proxy after bioturbation, seasonal bias, calibration bias and measurement noise                                                    \cr
proxy.bt.sb.sampYM.b           \tab proxy after bioturbation, seasonal bias, finite sampling and calibration bias                                                      \cr
proxy.bt.sb.sampYM.b.n         \tab proxy after bioturbation, seasonal bias, finite sampling, calibration bias and measurement noise                                   \cr
simulated.proxy            \tab final simulated proxy, this will be same as proxy.bt.sb.inf.b.n when n.samples = Inf, and proxy.bt.sb.sampYM.b.n when n.samples is finite
}
}
\description{
\code{ClimToProxyClim} simulates the creation of a proxy climate record
from a climate signal that is assumed to be true.

The following aspects of proxy creation are currently modelled.
\enumerate{
\item Seasonal bias in the encoding of a proxy due to the interaction between
climate seasonality and any seasonality in the life cycle of the organism
encoding the climate signal (e.g. Foraminifera for Mg/Ca ratios, or
phytoplankton for Alkenone unsaturation indices).
\item Bioturbation of the sediment archived proxy. For each requested
timepoint, the simulated proxy consists of a weighted mean of the climate
signal over a time window that is determined by the sediment accumulation
rate {sed.acc.rate} and the bioturbation depth {bio.depth} which defaults
to 0.1 m. The weights are given by the depth solution to an impulse
response function (Berger and Heath, 1968).
\item Aliasing of seasonal and inter-annual climate variation onto to
bioturbated (smoothed) signal.
For proxies measured on a small number of discrete particles both seasonal
and inter-annual climate variation is aliased into the proxy record. For
example, Foraminifera have a life-cycle of approximately 1 month, so they
record something like the mean temperature from a single month. If Mg/Ca is
measured on e.g. \code{n.samples} = 30 individuals, the measured proxy
signal is a mean of 30 distinct monthly mean temperatures and will thus be
a stochastic sample of the true mean climate.
\item Measurement noise/error is added as a pure Gaussian white noise
process with mean = 0, standard deviation = \code{meas.noise}.
\item Additionally, a random \emph{bias} can be added to each realisation of a
proxy record. Bias is simulated as a Gaussian random variable with mean =
0, standard deviation = \code{meas.bias}. The same randomly generated bias
value is applied to all timepoints in a simulated proxy record, when
multiple replicate proxies are generated ({n.replicates} > 1) each
replicate has a different bias applied.
}

\code{ClimToProxyClim} returns one or more replicates of the final simulated proxy
as well as several intermediate stages (see section \strong{Value} below).
}
