% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPFMs.R
\name{PlotPFMs}
\alias{PlotPFMs}
\title{Plot forward modelled sedimentary proxies}
\usage{
PlotPFMs(PFMs, colr.palette = NULL, alpha.palette = NULL, breaks,
  levl.labels)
}
\arguments{
\item{PFMs}{A dataframe of forward modelled proxies}

\item{colr.palette}{Colours for the proxy stages}

\item{alpha.palette}{Alpha levels for the proxy stages}

\item{breaks}{Proxy stages for legend, in order}

\item{levl.labels}{Labels for the proxy stages, in order}
}
\description{
Plot forward modelled sedimentary proxies
}
