% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPFMs.R
\name{PlotPFMs}
\alias{PlotPFMs}
\title{Plot forward modelled sedimentary proxies}
\usage{
PlotPFMs(PFMs, breaks = "default.breaks", max.replicates = 5,
  colr.palette = "default.colr.palette",
  alpha.palette = "default.alpha.palette",
  levl.labels = "default.levl.labels")
}
\arguments{
\item{PFMs}{A dataframe of forward modelled proxies}

\item{breaks}{Proxy stages for legend, in order}

\item{max.replicates}{Maximum number of replicates to plot at once}

\item{colr.palette}{Colours for the proxy stages}

\item{alpha.palette}{Alpha levels for the proxy stages}

\item{levl.labels}{Labels for the proxy stages, in order}
}
\description{
Plot forward modelled sedimentary proxies
}
