% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPFMs.R
\name{PlotPFMs}
\alias{PlotPFMs}
\title{Plot forward modelled sedimentary proxies}
\usage{
PlotPFMs(PFMs, plot.stages = c("default"), max.replicates = 5,
  colr.palette = "default", alpha.palette = "default",
  levl.labels = "default", facetted = FALSE)
}
\arguments{
\item{PFMs}{A dataframe of forward modelled proxies}

\item{plot.stages}{Proxy stages to be plotted, "default", "all", or a custom character vector}

\item{max.replicates}{Maximum number of replicates to plot at once}

\item{colr.palette}{Colours for the proxy stages}

\item{alpha.palette}{Alpha levels for the proxy stages}

\item{levl.labels}{Labels for the proxy stages}
}
\description{
Plot forward modelled sedimentary proxies
}
