% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedproxy.R
\docType{package}
\name{sedproxy}
\alias{sedproxy}
\title{sedproxy: Simulation of Sediment Archived Climate Proxy Records}
\description{
The sedproxy package provides functions to simulate sediment archived proxies
}
